pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logcraft"
version = "0.1.0"
description = "Craft CHANGELOG.md from git commit history with conventional-commit categorization"
readme = "README.md"
license = "MIT"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "gitpython>=3.1",
    "click>=8.0",
]

[project.scripts]
logcraft = "logcraft.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
